Assistant checkpoint: Fix leaderboard filter reset when adding friends
Assistant generated file changes:
- script.js: Preserve filter state when adding friend

---

User prompt:

i notice an issue. When I added a user to my friend list using the add friend button that is next to their name on the leaderboard, for some reason clicking it also reset the filters i had clicked -- can you tell why it did that and make it not do that?