Assistant checkpoint: Update Google sign-in to use email usernames
Assistant generated file changes:
- script.js: Update Google sign-in to use email username
Update Google sign-in to use email username

---

User prompt:

i just realized a problem that we are going to run into eventually when this scales with usernames and friends. When players login with Google login, the app is setting their username as their full name as it is on their google account. While this is somewhat convenient for them to login quickly, we will run into problems when multiple people have the same name and same username. Is there a solution around this? Like, can we make sure when accounts get made using google login, that their username defaults to save in our database as the username in their email? e.g. : mtpete33@gmail.com --> username: mtpete33 -- This would make it much easier to identify friends when there might be two Michael Peterson's in the search lookup... or do you have any other suggestion that would be a workaround for this? Also, I want to avoid having to make people make new accounts, or have to delete any accounts at this point, so could we have it just save their new usernames and overwrite their "name" the next time they log in?