Assistant checkpoint: Update friend adding logic to handle username changes
Assistant generated file changes:
- script.js: Update addFriend to handle username updates

---

User prompt:

i want to add a little clause to help alleviate discrepancy in the friend adding process into the database. the way i've altered the code, the way we save peoples 'username' is going to be starting to get changed when they log in next, so instead of using google's display name like Michael Peterson, it is going to start making their new 'username' in our database just the truncated username of their email like 'mtpete33' (from mtpete33@gmail.com) -- So I want to add a clause that when we save friends to a friend listing in firebase, which will log their uid and the username -- if the code encounters a duplicate uid but with a different username than before -- maybe we can just have the code overwrite the old username and use the newest one, if that makes sense? just as a failsafe...